+++ components/browser/state/src/main/java/mozilla/components/browser/state/reducer/TabListReducer.kt
@@ -21,21 +21,25 @@ internal object TabListReducer {
                 // Verify that tab doesn't already exist
                 requireUniqueTab(state, action.tab)
 
+                val isPrivate = action.tab.content.private
+                val tabList = if (isPrivate) state.privateTabs else state.normalTabs
+
                 val updatedTabList = if (action.tab.parentId != null) {
-                    val parentIndex = state.tabs.indexOfFirst { it.id == action.tab.parentId }
+                    val parentIndex = tabList.indexOfFirst { it.id == action.tab.parentId }
                     if (parentIndex == -1) {
                         throw IllegalArgumentException("The parent does not exist")
                     }
 
                     // Add the child tab next to its parent
                     val childIndex = parentIndex + 1
-                    state.tabs.subList(0, childIndex) + action.tab + state.tabs.subList(childIndex, state.tabs.size)
+                    tabList.subList(0, childIndex) + action.tab + tabList.subList(childIndex, tabList.size)
                 } else {
-                    state.tabs + action.tab
+                    tabList + action.tab
                 }
 
                 state.copy(
-                    tabs = updatedTabList,
+                    normalTabs = if (isPrivate) state.normalTabs else updatedTabList,
+                    privateTabs = if (isPrivate) updatedTabList else state.privateTabs,
                     selectedTabId = if (action.select || state.selectedTabId == null) {
                         action.tab.id
                     } else {
@@ -52,12 +56,10 @@ internal object TabListReducer {
                 }
 
                 state.copy(
-                    tabs = state.tabs + action.tabs,
-                    selectedTabId = if (state.selectedTabId == null) {
-                        action.tabs.find { tab -> !tab.content.private }?.id
-                    } else {
-                        state.selectedTabId
-                    }
+                    normalTabs = state.normalTabs + action.tabs.filterNot { it.content.private },
+                    privateTabs = state.privateTabs + action.tabs.filter { it.content.private },
+                    selectedTabId = state.selectedTabId
+                        ?: action.tabs.find { tab -> !tab.content.private }?.id
                 )
             }
 
@@ -71,8 +73,11 @@ internal object TabListReducer {
                 if (tabToRemove == null) {
                     state
                 } else {
+                    val isPrivate = tabToRemove.content.private
+                    val tabList = if (isPrivate) state.privateTabs else state.normalTabs
+
                     // Remove tab and update child tabs in case their parent was removed
-                    val updatedTabList = (state.tabs - tabToRemove).map {
+                    val updatedTabList = (tabList - tabToRemove).map {
                         if (it.parentId == tabToRemove.id) it.copy(parentId = tabToRemove.parentId) else it
                     }
 
@@ -81,15 +86,20 @@ internal object TabListReducer {
                         tabToRemove.parentId
                     } else if (state.selectedTabId == tabToRemove.id) {
                         // The selected tab was removed and we need to find a new one
-                        val previousIndex = state.tabs.indexOf(tabToRemove)
-                        findNewSelectedTabId(updatedTabList, tabToRemove.content.private, previousIndex)
+                        findNewSelectedTabId(
+                            state,
+                            updatedTabList,
+                            isPrivate = tabToRemove.content.private,
+                            previousIndex = tabList.indexOf(tabToRemove)
+                        )
                     } else {
                         // The selected tab is not affected and can stay the same
                         state.selectedTabId
                     }
 
                     state.copy(
-                        tabs = updatedTabList,
+                        normalTabs = if (isPrivate) state.normalTabs else updatedTabList,
+                        privateTabs = if (isPrivate) updatedTabList else state.privateTabs,
                         selectedTabId = updatedSelection
                     )
                 }
@@ -103,7 +113,8 @@ internal object TabListReducer {
                 // happen asynchronously we may already have a tab at this point (e.g. from an `Intent`) and so we
                 // pretend we restored the list of tabs before any tab was added.
                 state.copy(
-                    tabs = action.tabs + state.tabs,
+                    normalTabs = action.tabs.filterNot { it.content.private } + state.normalTabs,
+                    privateTabs = action.tabs.filter { it.content.private } + state.privateTabs,
                     selectedTabId = if (action.selectedTabId != null && state.selectedTabId == null) {
                         // We only want to update the selected tab if none has been already selected. Otherwise we may
                         // switch to a restored tab even though the user is already looking at an existing tab (e.g.
@@ -117,19 +128,19 @@ internal object TabListReducer {
 
             is TabListAction.RemoveAllTabsAction -> {
                 state.copy(
-                    tabs = emptyList(),
+                    normalTabs = emptyList(),
+                    privateTabs = emptyList(),
                     selectedTabId = null
                 )
             }
 
             is TabListAction.RemoveAllPrivateTabsAction -> {
                 val selectionAffected = state.selectedTab?.content?.private == true
-                val updatedTabs = state.tabs.filterNot { it.content.private }
                 state.copy(
-                    tabs = updatedTabs,
-                    selectedTabId = if (selectionAffected && updatedTabs.isNotEmpty()) {
+                    privateTabs = emptyList(),
+                    selectedTabId = if (selectionAffected && state.normalTabs.isNotEmpty()) {
                         // If the selection is affected, select the last normal tab, if available.
-                        updatedTabs.last().id
+                        state.normalTabs.last().id
                     } else {
                         state.selectedTabId
                     }
@@ -138,9 +149,8 @@ internal object TabListReducer {
 
             is TabListAction.RemoveAllNormalTabsAction -> {
                 val selectionAffected = state.selectedTab?.content?.private == false
-                val updatedTabs = state.tabs.filter { it.content.private }
                 state.copy(
-                    tabs = updatedTabs,
+                    normalTabs = emptyList(),
                     selectedTabId = if (selectionAffected) {
                         // If the selection is affected, we'll set it to null as there's no
                         // normal tab left and NO private tab should get selected instead.
@@ -158,32 +168,45 @@ internal object TabListReducer {
  * Find a new selected tab and return its id after the tab at [previousIndex] was removed.
  */
 private fun findNewSelectedTabId(
-    tabs: List<TabSessionState>,
+    state: BrowserState,
+    newTabs: List<TabSessionState>,
     isPrivate: Boolean,
     previousIndex: Int
 ): String? {
-    if (tabs.isEmpty()) {
+    val allTabs: List<TabSessionState>
+    val previousIndexAll: Int
+    if (isPrivate) {
+        // Can switch from private tabs to normal tabs
+        allTabs = state.normalTabs + newTabs
+        previousIndexAll = state.normalTabs.size + previousIndex
+    } else {
+        // Cannot switch from normal tabs to private tabs
+        allTabs = newTabs
+        previousIndexAll = previousIndex
+    }
+
+    if (allTabs.isEmpty()) {
         // There's no tab left to select.
         return null
     }
 
-    val predicate: (TabSessionState) -> Boolean = { tab -> tab.content.private == isPrivate }
-
     // If the previous index is still a valid index and if this is a private/normal tab we are looking for then
     // let's use the tab at the same index.
-    if (previousIndex <= tabs.lastIndex && predicate(tabs[previousIndex])) {
-        return tabs[previousIndex].id
+    if (previousIndex in newTabs.indices) {
+        return newTabs[previousIndex].id
     }
 
     // Find a tab that matches the predicate and is nearby the tab that was selected previously
-    val nearbyTab = findNearbyTab(tabs, previousIndex, predicate)
+    val nearbyTab = findNearbyTab(allTabs, previousIndexAll) { tab ->
+        tab.content.private == isPrivate
+    }
 
     return when {
         // We found a nearby tab, let's select it.
         nearbyTab != null -> nearbyTab.id
 
         // If there's no private tab to select anymore then just select the last regular tab
-        isPrivate -> tabs.last().id
+        isPrivate -> state.normalTabs.lastOrNull()?.id
 
         // Removing the last normal tab should NOT cause a private tab to be selected
         else -> null
@@ -224,7 +247,7 @@ private fun findNearbyTab(
  * @param tab the [TabSessionState] to check.
  */
 private fun requireUniqueTab(state: BrowserState, tab: TabSessionState) {
-    require(state.tabs.find { it.id == tab.id } == null) {
+    require(state.tabs.none { it.id == tab.id }) {
         "Tab with same ID already exists"
     }
 }