+++ components/browser/state/src/test/java/mozilla/components/browser/state/action/EngineActionTest.kt
@@ -29,7 +29,7 @@ class EngineActionTest {
 
         store = BrowserStore(
             initialState = BrowserState(
-                tabs = listOf(tab)
+                normalTabs = listOf(tab)
             )
         )
     }